allows the `customText`-pipe to take sone replacements parameters, in order to replace stuff in the customText.
So we dont need two customTexts in a scenario like this:
```
<div>
  {{ 'Do you want to delete the' | customtext:'deletion_message_part1':cts.updateCount}}
  {{numberOfItems}}
  {{ ' files or move them into trash?' | customtext:'deletion_message_part2':cts.updateCount}}
</div>
```

This could no be written as

```
<div>{{ 'Do you want to delete the %s files or move them into trash?' | customtext:'deletion_message':cts.updateCount:numberOfItems}}
```

I chose `%s` as replacement token in dependence on the `printf`-function well known from various programming languages. Since the parameter can only be a string (due to the nature of the pipe), it did not made sense to implement the other tokens from `sprintf` like `%i`.